requirements/base.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/base.in -o requirements/base.txt
asgiref==3.8.1
    # via django
django==5.1.3
    # via
    #   -r requirements/base.in
    #   django-filter
    #   djangorestframework
django-filter==24.3
    # via -r requirements/base.in
djangorestframework==3.15.2
    # via -r requirements/base.in
sqlparse==0.5.2
    # via django
